Align Uno responses with Pico protocol
Correct analog availability records, preserve binary LF bytes, restore six-decimal PWM formatting, and add golden wire-format coverage.

+++ source/visa/analog.cpp
@@ -34,7 +34,7 @@ namespace Visa {
         uint16_t val;
         gPlatform.IO.Analog.GetInput(channel, &val);
 
-        gPlatform.IO.Printf("%d\n", val);
+        gPlatform.IO.Printf("%u\n", static_cast<unsigned>(val));
 
         return QueryResult::Success;
     }
